requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
azure-cognitiveservices-speech>=1.31.0 